+++ apps/ledger-box/src/modules/wallet-settings/wallet-settings-danger-zone/index.ts
@@ -1,0 +1,2 @@
+export { WalletSettingsDangerZone } from '#/modules/wallet-settings/wallet-settings-danger-zone/wallet-settings-danger-zone';
+export { WalletSettingsDangerZoneRoute } from '#/modules/wallet-settings/wallet-settings-danger-zone/wallet-settings-danger-zone-route';