+++ apps/ledger-box/src/modules/wallet-settings/wallet-settings-general/index.ts
@@ -1,0 +1,2 @@
+export { WalletSettingsGeneral } from '#/modules/wallet-settings/wallet-settings-general/wallet-settings-general';
+export { WalletSettingsGeneralRoute } from '#/modules/wallet-settings/wallet-settings-general/wallet-settings-general-route';