+++ apps/ledger-box/src/modules/wallet-settings/wallet-settings-general/wallet-settings-general.tsx
@@ -6,7 +6,7 @@ import { Field, FieldError, FieldGroup, FieldLabel } from '@vhnam/ui/components/
 import { Input } from '@vhnam/ui/components/input';
 import { Spinner } from '@vhnam/ui/components/spinner';
 
-import { useWalletSettingsGeneralActions } from '#/modules/wallets/wallet-settings-general/wallet-settings-general.actions';
+import { useWalletSettingsGeneralActions } from '#/modules/wallet-settings/wallet-settings-general/wallet-settings-general.actions';
 import type { WalletDto } from '#/queries/wallets/wallet.dto';
 
 type WalletSettingsGeneralProps = {