+++ apps/ledger-box/src/modules/wallet-settings/wallet-settings-members/index.ts
@@ -1,0 +1,2 @@
+export { WalletSettingsMembers } from '#/modules/wallet-settings/wallet-settings-members/wallet-settings-members';
+export { WalletSettingsMembersRoute } from '#/modules/wallet-settings/wallet-settings-members/wallet-settings-members-route';