+++ apps/ledger-box/src/modules/wallets/wallet-settings-members/index.ts
@@ -1,1 +1,0 @@
-export { WalletSettingsMembers } from '#/modules/wallets/wallet-settings-members/wallet-settings-members';